server/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-JWT-Extended==4.7.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
